requirements.txt
python-dotenv
spaCy
tiktoken
Pillow
openai
pyttsx3
SpeechRecognition
pyaudio
word2number
python-dotenv